+++ packages/wonder-blocks-modal/src/components/__docs__/modal-dialog.stories.js
@@ -46,10 +46,26 @@ const customViewports = {
 export default {
     title: "Modal/Building Blocks/ModalDialog",
     component: ModalDialog,
+    decorators: [
+        (Story: any): React.Element<typeof View> => (
+            <View style={styles.example}>
+                <Story />
+            </View>
+        ),
+    ],
     parameters: {
         componentSubtitle: ((
             <ComponentInfo name={name} version={version} />
         ): any),
+        docs: {
+            description: {
+                component: null,
+            },
+            source: {
+                // See https://github.com/storybookjs/storybook/issues/12596
+                excludeDecorators: true,
+            },
+        },
         viewport: {
             viewports: customViewports,
             defaultViewport: "desktop",
@@ -245,6 +261,10 @@ WithLauncher.parameters = {
 };
 
 const styles = StyleSheet.create({
+    example: {
+        alignItems: "center",
+        justifyContent: "center",
+    },
     modalPositioner: {
         // Checkerboard background
         backgroundImage:
@@ -263,7 +283,8 @@ const styles = StyleSheet.create({
         bottom: 0,
     },
     previewSizer: {
-        height: 600,
+        minHeight: 600,
+        width: "100%",
     },
     row: {
         flexDirection: "row",